refactored entity attribute registry
