+++ src/main/java/net/quantumgoddess/soulscraft/entity/attribute/SoulsCraftEntityAttributes.java
@@ -1,0 +1,5 @@
+package net.quantumgoddess.soulscraft.entity.attribute;
+
+public class SoulsCraftEntityAttributes {
+
+}
